Evolve4
App Name is "Evolve4
This file or set of files can servee as a branch or braches of the main app name DT-BDLM, or its own app. Both use the 8 A's of Transformation. 
These app builds are intended to be a meditation app, which has a picture or set of Pictures which depict 8 A's of Thranformation
The 8 A's are 1. Awareness; 2. Acceptance; 3. Assurance; 4. Adoration; 5. Association; 6. Assignment; 7. Accountability and 8. Assimilation
 
my name is Dennis L. Waters, Sr.